Move manifest nodes to dbt/artifacts (#9538)
