+++ .changes/unreleased/Under the Hood-20240207-122342.yaml
@@ -1,0 +1,6 @@
+kind: Under the Hood
+body: Move manifest nodes to artifacts
+time: 2024-02-07T12:23:42.909049-05:00
+custom:
+  Author: gshank
+  Issue: "9388"
