--- core/dbt/artifacts/resources/v1/docs.py
@@ -1,9 +1,0 @@
-from dataclasses import dataclass
-from dbt_common.dataclass_schema import dbtClassMixin
-from typing import Optional
-
-
-@dataclass
-class Docs(dbtClassMixin):
-    show: bool = True
-    node_color: Optional[str] = None
