+++ core/dbt/contracts/files.py
@@ -1,4 +1,3 @@
-import hashlib
 import os
 from dataclasses import dataclass, field
 
@@ -7,6 +6,7 @@ from typing import List, Optional, Union, Dict, Any
 
 from dbt.constants import MAXIMUM_SEED_SIZE
 from dbt_common.dataclass_schema import dbtClassMixin, StrEnum
+from dbt.artifacts.resources.base import FileHash
 
 from .util import SourceKey
 
@@ -68,46 +68,6 @@ class FilePath(dbtClassMixin):
     def seed_too_large(self) -> bool:
         """Return whether the file this represents is over the seed size limit"""
         return os.stat(self.full_path).st_size > MAXIMUM_SEED_SIZE
-
-
-@dataclass
-class FileHash(dbtClassMixin):
-    name: str  # the hash type name
-    checksum: str  # the hashlib.hash_type().hexdigest() of the file contents
-
-    @classmethod
-    def empty(cls):
-        return FileHash(name="none", checksum="")
-
-    @classmethod
-    def path(cls, path: str):
-        return FileHash(name="path", checksum=path)
-
-    def __eq__(self, other):
-        if not isinstance(other, FileHash):
-            return NotImplemented
-
-        if self.name == "none" or self.name != other.name:
-            return False
-
-        return self.checksum == other.checksum
-
-    def compare(self, contents: str) -> bool:
-        """Compare the file contents with the given hash"""
-        if self.name == "none":
-            return False
-
-        return self.from_contents(contents, name=self.name) == self.checksum
-
-    @classmethod
-    def from_contents(cls, contents: str, name="sha256") -> "FileHash":
-        """Create a file hash from the given file contents. The hash is always
-        the utf-8 encoding of the contents given, because dbt only reads files
-        as utf-8.
-        """
-        data = contents.encode("utf-8")
-        checksum = hashlib.new(name, data).hexdigest()
-        return cls(name=name, checksum=checksum)
 
 
 @dataclass
