+++ core/dbt/contracts/graph/nodes.py
@@ -19,14 +19,9 @@ from typing import (
 )
 
 from dbt import deprecations
-from dbt_common.contracts.constraints import (
-    ConstraintType,
-    ModelLevelConstraint,
-)
-from dbt_common.dataclass_schema import dbtClassMixin
+from dbt_common.contracts.constraints import ConstraintType
 
 from dbt_common.clients.system import write_file
-from dbt.contracts.files import FileHash
 from dbt.contracts.graph.unparsed import (
     HasYamlMetadata,
     TestDef,
@@ -37,6 +32,11 @@ from dbt.contracts.graph.unparsed import (
     UnitTestInputFixture,
     UnitTestOutputFixture,
     UnitTestNodeVersions,
+)
+from dbt.contracts.graph.model_config import (
+    UnitTestNodeConfig,
+    UnitTestConfig,
+    EmptySnapshotConfig,
 )
 from dbt.contracts.graph.node_args import ModelNodeArgs
 from dbt_common.events.functions import warn_or_error
@@ -57,25 +57,12 @@ from dbt.node_types import (
     VERSIONED_NODE_TYPES,
 )
 
-from .model_config import (
-    NodeConfig,
-    ModelConfig,
-    SeedConfig,
-    TestConfig,
-    EmptySnapshotConfig,
-    SnapshotConfig,
-    UnitTestConfig,
-    UnitTestNodeConfig,
-)
 
 from dbt.artifacts.resources import (
     BaseResource,
-    ColumnInfo as ColumnInfoResource,
     DependsOn,
     Docs,
     Exposure as ExposureResource,
-    HasRelationMetadata as HasRelationMetadataResource,
-    MacroDependsOn,
     MacroArgument,
     Documentation as DocumentationResource,
     Macro as MacroResource,
@@ -83,10 +70,26 @@ from dbt.artifacts.resources import (
     NodeVersion,
     Group as GroupResource,
     GraphResource,
-    Quoting as QuotingResource,
-    RefArgs as RefArgsResource,
     SavedQuery as SavedQueryResource,
     SemanticModel as SemanticModelResource,
+    ParsedResourceMandatory,
+    ParsedResource,
+    CompiledResource,
+    HasRelationMetadata as HasRelationMetadataResource,
+    FileHash,
+    NodeConfig,
+    ColumnInfo,
+    InjectedCTE,
+    Analysis as AnalysisResource,
+    HookNode as HookNodeResource,
+    Model as ModelResource,
+    ModelConfig,
+    SqlOperation as SqlOperationResource,
+    Seed as SeedResource,
+    SingularTest as SingularTestResource,
+    GenericTest as GenericTestResource,
+    Snapshot as SnapshotResource,
+    Quoting as QuotingResource,
     SourceDefinition as SourceDefinitionResource,
 )
 
@@ -180,35 +183,30 @@ class GraphNode(GraphResource, BaseNode):
 
 
 @dataclass
-class Contract(dbtClassMixin):
-    enforced: bool = False
-    alias_types: bool = True
-    checksum: Optional[str] = None
+class HasRelationMetadata(HasRelationMetadataResource):
+    @classmethod
+    def __pre_deserialize__(cls, data):
+        data = super().__pre_deserialize__(data)
+        if "database" not in data:
+            data["database"] = None
+        return data
+
+    @property
+    def quoting_dict(self) -> Dict[str, bool]:
+        if hasattr(self, "quoting"):
+            return self.quoting.to_dict(omit_none=True)
+        else:
+            return {}
 
 
 @dataclass
-class DeferRelation(HasRelationMetadataResource):
-    alias: str
-    relation_name: Optional[str]
-
-    @property
-    def identifier(self):
-        return self.alias
-
-
-@dataclass
-class ParsedNodeMandatory(GraphNode, HasRelationMetadataResource):
-    alias: str
-    checksum: FileHash
-    config: NodeConfig = field(default_factory=NodeConfig)
-
-    @property
-    def identifier(self):
-        return self.alias
+class ParsedNodeMandatory(ParsedResourceMandatory, GraphNode, HasRelationMetadata):
+    pass
 
 
 # This needs to be in all ManifestNodes and also in SourceDefinition,
-# because of "source freshness"
+# because of "source freshness". Should not be in artifacts, because we
+# don't write out _event_status.
 @dataclass
 class NodeInfoMixin:
     _event_status: Dict[str, Any] = field(default_factory=dict)
@@ -244,22 +242,7 @@ class NodeInfoMixin:
 
 
 @dataclass
-class ParsedNode(NodeInfoMixin, ParsedNodeMandatory, SerializableType):
-    tags: List[str] = field(default_factory=list)
-    description: str = field(default="")
-    columns: Dict[str, ColumnInfoResource] = field(default_factory=dict)
-    meta: Dict[str, Any] = field(default_factory=dict)
-    group: Optional[str] = None
-    docs: Docs = field(default_factory=Docs)
-    patch_path: Optional[str] = None
-    build_path: Optional[str] = None
-    deferred: bool = False
-    unrendered_config: Dict[str, Any] = field(default_factory=dict)
-    created_at: float = field(default_factory=lambda: time.time())
-    config_call_dict: Dict[str, Any] = field(default_factory=dict)
-    relation_name: Optional[str] = None
-    raw_code: str = ""
-
+class ParsedNode(ParsedResource, NodeInfoMixin, ParsedNodeMandatory, SerializableType):
     def get_target_write_path(self, target_path: str, subdirectory: str):
         # This is called for both the "compiled" subdirectory of "target" and the "run" subdirectory
         if os.path.basename(self.path) == os.path.basename(self.original_file_path):
@@ -299,8 +282,6 @@ class ParsedNode(NodeInfoMixin, ParsedNodeMandatory, SerializableType):
             return AnalysisNode.from_dict(dct)
         elif resource_type == "seed":
             return SeedNode.from_dict(dct)
-        elif resource_type == "rpc":
-            return RPCNode.from_dict(dct)
         elif resource_type == "sql":
             return SqlNode.from_dict(dct)
         elif resource_type == "test":
@@ -398,30 +379,9 @@ class ParsedNode(NodeInfoMixin, ParsedNodeMandatory, SerializableType):
 
 
 @dataclass
-class InjectedCTE(dbtClassMixin):
-    """Used in CompiledNodes as part of ephemeral model processing"""
-
-    id: str
-    sql: str
-
-
-@dataclass
-class CompiledNode(ParsedNode):
+class CompiledNode(CompiledResource, ParsedNode):
     """Contains attributes necessary for SQL files and nodes with refs, sources, etc,
     so all ManifestNodes except SeedNode."""
-
-    language: str = "sql"
-    refs: List[RefArgsResource] = field(default_factory=list)
-    sources: List[List[str]] = field(default_factory=list)
-    metrics: List[List[str]] = field(default_factory=list)
-    depends_on: DependsOn = field(default_factory=DependsOn)
-    compiled_path: Optional[str] = None
-    compiled: bool = False
-    compiled_code: Optional[str] = None
-    extra_ctes_injected: bool = False
-    extra_ctes: List[InjectedCTE] = field(default_factory=list)
-    _pre_injected_sql: Optional[str] = None
-    contract: Contract = field(default_factory=Contract)
 
     @property
     def empty(self):
@@ -469,27 +429,17 @@ class CompiledNode(ParsedNode):
 
 
 @dataclass
-class AnalysisNode(CompiledNode):
-    resource_type: Literal[NodeType.Analysis]
+class AnalysisNode(AnalysisResource, CompiledNode):
+    pass
 
 
 @dataclass
-class HookNode(CompiledNode):
-    resource_type: Literal[NodeType.Operation]
-    index: Optional[int] = None
+class HookNode(HookNodeResource, CompiledNode):
+    pass
 
 
 @dataclass
-class ModelNode(CompiledNode):
-    resource_type: Literal[NodeType.Model]
-    access: AccessType = AccessType.Protected
-    config: ModelConfig = field(default_factory=ModelConfig)
-    constraints: List[ModelLevelConstraint] = field(default_factory=list)
-    version: Optional[NodeVersion] = None
-    latest_version: Optional[NodeVersion] = None
-    deprecation_date: Optional[datetime] = None
-    defer_relation: Optional[DeferRelation] = None
-
+class ModelNode(ModelResource, CompiledNode):
     @classmethod
     def from_args(cls, args: ModelNodeArgs) -> "ModelNode":
         unique_id = args.unique_id
@@ -766,15 +716,9 @@ class ModelNode(CompiledNode):
         return False
 
 
-# TODO: rm?
 @dataclass
-class RPCNode(CompiledNode):
-    resource_type: Literal[NodeType.RPCCall]
-
-
-@dataclass
-class SqlNode(CompiledNode):
-    resource_type: Literal[NodeType.SqlOperation]
+class SqlNode(SqlOperationResource, CompiledNode):
+    pass
 
 
 # ====================================
@@ -783,15 +727,7 @@ class SqlNode(CompiledNode):
 
 
 @dataclass
-class SeedNode(ParsedNode):  # No SQLDefaults!
-    resource_type: Literal[NodeType.Seed]
-    config: SeedConfig = field(default_factory=SeedConfig)
-    # seeds need the root_path because the contents are not loaded initially
-    # and we need the root_path to load the seed later
-    root_path: Optional[str] = None
-    depends_on: MacroDependsOn = field(default_factory=MacroDependsOn)
-    defer_relation: Optional[DeferRelation] = None
-
+class SeedNode(SeedResource, ParsedNode):  # No SQLDefaults!
     def same_seeds(self, other: "SeedNode") -> bool:
         # for seeds, we check the hashes. If the hashes are different types,
         # no match. If the hashes are both the same 'path', log a warning and
@@ -909,12 +845,7 @@ class TestShouldStoreFailures:
 
 
 @dataclass
-class SingularTestNode(TestShouldStoreFailures, CompiledNode):
-    resource_type: Literal[NodeType.Test]
-    # Was not able to make mypy happy and keep the code working. We need to
-    # refactor the various configs.
-    config: TestConfig = field(default_factory=TestConfig)  # type: ignore
-
+class SingularTestNode(SingularTestResource, TestShouldStoreFailures, CompiledNode):
     @property
     def test_node_type(self):
         return "singular"
@@ -926,34 +857,7 @@ class SingularTestNode(TestShouldStoreFailures, CompiledNode):
 
 
 @dataclass
-class TestMetadata(dbtClassMixin):
-    __test__ = False
-
-    name: str
-    # kwargs are the args that are left in the test builder after
-    # removing configs. They are set from the test builder when
-    # the test node is created.
-    kwargs: Dict[str, Any] = field(default_factory=dict)
-    namespace: Optional[str] = None
-
-
-# This has to be separated out because it has no default and so
-# has to be included as a superclass, not an attribute
-@dataclass
-class HasTestMetadata(dbtClassMixin):
-    test_metadata: TestMetadata
-
-
-@dataclass
-class GenericTestNode(TestShouldStoreFailures, CompiledNode, HasTestMetadata):
-    resource_type: Literal[NodeType.Test]
-    column_name: Optional[str] = None
-    file_key_name: Optional[str] = None
-    # Was not able to make mypy happy and keep the code working. We need to
-    # refactor the various configs.
-    config: TestConfig = field(default_factory=TestConfig)  # type: ignore
-    attached_node: Optional[str] = None
-
+class GenericTestNode(GenericTestResource, TestShouldStoreFailures, CompiledNode):
     def same_contents(self, other, adapter_type: Optional[str]) -> bool:
         if other is None:
             return False
@@ -1072,10 +976,8 @@ class IntermediateSnapshotNode(CompiledNode):
 
 
 @dataclass
-class SnapshotNode(CompiledNode):
-    resource_type: Literal[NodeType.Snapshot]
-    config: SnapshotConfig
-    defer_relation: Optional[DeferRelation] = None
+class SnapshotNode(SnapshotResource, CompiledNode):
+    pass
 
 
 # ====================================
@@ -1220,7 +1122,7 @@ class SourceDefinition(
     NodeInfoMixin,
     GraphNode,
     SourceDefinitionResource,
-    HasRelationMetadataResource,
+    HasRelationMetadata,
 ):
     @classmethod
     def resource_class(cls) -> Type[SourceDefinitionResource]:
@@ -1618,7 +1520,7 @@ class ParsedPatch(HasYamlMetadata):
 # may be empty.
 @dataclass
 class ParsedNodePatch(ParsedPatch):
-    columns: Dict[str, ColumnInfoResource]
+    columns: Dict[str, ColumnInfo]
     access: Optional[str]
     version: Optional[NodeVersion]
     latest_version: Optional[NodeVersion]
@@ -1643,7 +1545,6 @@ ManifestSQLNode = Union[
     SingularTestNode,
     HookNode,
     ModelNode,
-    RPCNode,
     SqlNode,
     GenericTestNode,
     SnapshotNode,
