+++ core/dbt/parser/manifest.py
@@ -79,7 +79,7 @@ from dbt.context.docs import generate_runtime_docs_context
 from dbt.context.macro_resolver import MacroResolver, TestMacroNamespace
 from dbt.context.configured import generate_macro_context
 from dbt.context.providers import ParseProvider, generate_runtime_macro_context
-from dbt.contracts.files import FileHash, ParseFileType, SchemaSourceFile
+from dbt.contracts.files import ParseFileType, SchemaSourceFile
 from dbt.parser.read_files import (
     ReadFilesFromFileSystem,
     load_source_file,
@@ -107,7 +107,7 @@ from dbt.contracts.graph.nodes import (
     ResultNode,
     ModelNode,
 )
-from dbt.artifacts.resources import NodeRelation, NodeVersion
+from dbt.artifacts.resources import NodeRelation, NodeVersion, FileHash
 from dbt.artifacts.schemas.base import Writable
 from dbt.exceptions import (
     TargetNotFoundError,
